Move all use over to BOOST_SAFE_NUMBERS_THROW_EXCEPTION
